Address review comments
- Use `ImmutableArray<T>`.
- Fix typos.
- Use correct return type.

+++ tests/AdventOfCode.Tests/Puzzles/Y2025/Day10Tests.cs
@@ -17,7 +17,7 @@ public sealed class Day10Tests(ITestOutputHelper outputHelper) : PuzzleTest(outp
         ];
 
         // Act
-        (int actualIndicator, long actualJoltage) = Day10.GetMinimumButtonPresses(values, TestContext.Current.CancellationToken);
+        (int actualIndicator, int actualJoltage) = Day10.GetMinimumButtonPresses(values, TestContext.Current.CancellationToken);
 
         // Assert
         actualIndicator.ShouldBe(7);
